bitcoin offchain (#12)
* bitcoin API

* bcash API

* bitcoin API

* litecoin API

* ethereum API

* xrp API

* xlm API

* return pointer

* vet API

* ledger API

* fixed import packages

* testing code

* validate

* testing code

* kms, record APIs

* bitcoin transaction builder

* bitcoin transaction

* bcash transaction

* private keys, addresses of btc, ltc and bch

* test

* wallet

* remove unused files

* ltc wallet

* ethereum transaction

* ethereum transaction

* bitcoin offchain

* litecoin offchain

* litecoin offchain

* bcash offchain

* eth offchain

* vet wallet

* eth wallet

* vet tx

* wbtc

* wallet and private key

* wallet and private key

* chain id

* fixed errors

+++ model/response/offchain/address.go
@@ -20,7 +20,7 @@ type Address struct {
 	 * @type {number}
 	 * @memberof Address
 	 */
-	DerivationKey uint64
+	DerivationKey uint32
 	/**
 	 * Extended public key to derive address from. In case of XRP, this is account address,
 	 * since address is defined as DestinationTag, which is address field. In case of XLM, this is account address, since address is defined as message, which is address field.
